Move Card to Components and rename to SideBar
